(incomplete) re-organized directories and testing scripts. (#225)
* re-organized directories and testing scripts. Changes amply_tests.py to use unittests.

# Conflicts:
#	pulp/pulp.cfg.win

* changed README to rst format. Added Sphinx helper make files.

# Conflicts:
#	setup.py

* closes #55
changed README to rst format.

# Conflicts:
#	README.rst

* fixes #222

* fixes #223

# Conflicts:
#	examples/wedding_initial.py

* fixes #203

* changed reference to README.rst

* modifications to change testing of library.

* corrected problem when reusing the same solver object in all tests by copying the solver. Tests passe now (except COINMP_DLL).

* edited tests so they all run together from run_tests.py. Added Franco Peschiera to AUTHORS.

* added examples and doc tests to run_tests.py

* cleaned old files.

* taken out references to pipenv for now. we keep like before with pip install .

* see if it fixes python 2.X errors and detects tests fails in python 3.X

* compatibility with pythons 2.X and return the correct status when tests fail.

* edited some tests, added workaround for GUROBI libraries loading. Changed permissions to internal solvers.

* more issues with 2.6. Fixed imports for amply, issue with 2.7 and, finally, the issue with finding the examples directory.

* solved compatibility with 2.7 for examples and made a wide exception catch in CPLEX to catch incompatibility with 2.7. Now we show the exception in the PulpSolverError.

* take out python2.6 from travis
