+++ AUTHORS
@@ -1,3 +1,4 @@
 Roy, J.S
 Mitchell, Stuart A
 Christophe-Marie Duquesne (PYGLPK and YAPOSIB bindings)
+Franco Peschiera
