--- docs.cfg
@@ -1,11 +1,0 @@
-[buildout]
-extends = buildout.cfg
-parts += sphinx 
-
-[sphinx]
-recipe = collective.recipe.sphinxbuilder
-source = ${buildout:directory}/doc/source
-build = ${buildout:directory}/doc
-interpreter = ${buildout:directory}/bin/pythonpulp
-outputs = html
-
