+++ examples/test3.py
@@ -8,6 +8,7 @@
 # The thermal units have a proportional cost and a startup cost.
 # The hydro unit has an initial storage.
 
+from __future__ import print_function
 from pulp import *
 from math import *
