+++ pulp/__init__.py
@@ -35,5 +35,3 @@ from .pulp import *
 from .amply import *
 __doc__ = pulp.__doc__
 __version__ = VERSION
-
-from . import tests
