+++ pulp/amply.py
@@ -691,7 +691,7 @@ else:
     param_record = param_slice_record | plain_data_record | tabular_record | \
             tr_tabular_record | Suppress(":=")
 
-    param_default = Optional("default" + data.setResultsName('default'))
+    param_default = Optional("default" + number.setResultsName('default'))
 
     param_stmt = "param" + symbol.setResultsName('name') + param_default + \
             Group(OneOrMore(param_record)).setResultsName('records') + END
@@ -705,6 +705,8 @@ else:
     param_def_stmt = "param" + symbol + Optional(subscript_domain) + \
             param_default + END
     param_def_stmt.setParseAction(ParamDefStmt)
+
+    # comment = "#" + SkipTo(lineEnd)
 
     stmts = set_stmt | set_def_stmt | param_stmt | param_def_stmt | \
             param_tabbing_stmt
