--- setup.cfg
@@ -1,7 +1,0 @@
-[build_sphinx]
-source-dir = doc/source
-build-dir  = build/doc
-all_files  = 1
-
-[upload_sphinx]
-upload-dir = doc/build/html
