+++ setup.py
@@ -4,8 +4,6 @@ Setup script for PuLP added by Stuart Mitchell 2007
 Copyright 2007 Stuart Mitchell
 """
 import sys
-from ez_setup import use_setuptools
-use_setuptools()
 from setuptools import setup
 
 readme_name = 'README.rst'
@@ -15,7 +13,7 @@ Description = open(readme_name).read()
 
 # read the version number safely from the constants.py file
 version_dict = {}
-exec(open('src/pulp/constants.py').read(), version_dict)
+exec(open('pulp/constants.py').read(), version_dict)
 VERSION = version_dict['VERSION']
 
 with open(readme_name, "r") as fh:
@@ -41,7 +39,6 @@ setup(name="PuLP",
                      'Topic :: Scientific/Engineering :: Mathematics',
       ],
       #ext_modules = [pulpCOIN],
-      package_dir={'':'src'},
       #need the cbc directories here as the executable bit is set
       packages = ['pulp',
       'pulp.solverdir',
